app/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
tensorflow==2.18.0
seaborn==0.13.2
opencv-python==4.11.0
Pillow==11.1.0
tensorflow-hub==0.16.1
streamlit==1.35.0